requirements.txt
tf-nightly
gin-config
tensorflow_hub
tensorflow_addons
tqdm
pandas
scikit-learn
seaborn
matplotlib==3.1.0
google-cloud-storage
spacy
emoji
unidecode
